整理 pytest 报错文档并稳定单测；补齐类型桩、收敛 mypy 告警，修复数据源与配置问题
- 文档与可观测性
  - 系统化记录 pytest 全量/离线执行的失败用例与原因，覆盖 AmazingData `SystemExit`、网络依赖、缓存/数据库连接等情境（`记录 pytest 全量执行失败原因`、`更新 pytest 报错记录`、`docs: 记录 mypy 剩余问题`、`chore: document offline unit test run`）。
  - 收敛 mypy 检查范围，移除不必要的 overrides，并补充报告与环境探测脚本（`收敛 mypy 检查范围并补齐类型桩`、`移除 mypy overrides 并更新报告`、`完善 mypy 环境探测`、`chore: add stubs and update mypy report`）。

- 测试稳定性与修复
  - 修复 AmazingData 相关问题：登录补丁路径、历史 K 线默认周期推断、市场概览兜底处理（`修复 AmazingData 登录补丁路径并 补充单测`、`修复 AmazingData 历史K线默认周期推断并更新测试报告`、`完善测试模式启用逻辑并补充市场概览兜底`）。
  - 数据源管理与并发：隔离数据源管理桩、修复 API 测试、修复数据源工厂清理死锁（`隔离数据源管理桩并修复API测试`、`修复数据源工厂清理死锁并补充pytest记录`）。
  - 配置与依赖：启用数据源配置并自动补齐开发配置文件；使用 `uv` 同步依赖；调整市场数据 API 响应头的数据源标识（`修复数据源配置启用并自动补齐开发配置文件`、`使用 uv 同步依赖并更新 pytest 报告`、`调整市场数据API响应头数据源标识`）。
  - 单测体验：完善 `MockDataProvider` 测试并记录最新 pytest 状态（`完善 MockDataProvider 测试并记录最新 pytest 状态`）。
  - 其他修复：缓存与数据库连接测试警告修复（`修复缓存与数据库连接测试警告`）。

- 类型系统与静态检查（mypy）
  - 大量补齐与完善类型桩，覆盖 AmazingData、QMT、AkShare、Redis、数据库、领域实体、代理配置等（`完善AmazingData类型桩与领域实体 兜底`、`分类治理：补齐Redis桩与AkShare/QMT类型修复`、`完善数据库与外部依赖的类型声明`、`完善代理配置类型并补齐数据库桩`）。
  - 收敛第三方与内部接口的 mypy 告警、判空与兜底处理（`收敛AmazingData与QMT的mypy告警`、`完善QMT与AmazingData的类型判空`、`完善类型桩并修复缓存与市场接口类型告警`）。
  - WebUI 与后端接口适配 mypy 要求（`Refine WebUI data endpoints for mypy compliance`）。
  - AkShare 特殊修复与依赖导入兜底（`修复 AkShare 类型检查问题并补充 依赖导入兜底`）。

- 影响范围
  - 代码涉及数据源集成（AmazingData、QMT、AkShare）、缓存/数据库、WebUI 接口、类型桩与 mypy 配置、测试框架与离线执行文档。
  - 对外 API 行为：市场数据 API 响应头中包含数据源标识；其余改动主要为类型与稳定性增强，不预期破坏性变更。

- 升级与运行建议
  - 使用 `uv` 进行依赖同步，确保本地与 CI 环境一致。
  - 启用测试模式或离线运行时，参考新增文档说明进行配置文件自动补齐与依赖隔离。
  - 首次运行建议全量执行 `pytest -q`，若网络受限可按文档选择性跳过外部依赖用例。

+++ deepsearch/infrastructure/providers/proxy/validator.py
@@ -44,7 +44,7 @@ class ProxyValidator:
     ]
     ANONYMITY_CHECK_URL = "http://httpbin.org/headers"
 
-    def __init__(self, timeout: int = 10) -> None:
+    def __init__(self, timeout: float = 10) -> None:
         self.timeout = timeout
         self._session: Optional[aiohttp.ClientSession] = None
